[#422] - Migrate CLI to Shared+CLI modules.
