+++ .github/CODEOWNERS
@@ -4,6 +4,8 @@
 
 /cli/ @aidanm3341 @lbulanti-ms @willosborne @grahampacker-ms @jpgough-ms @rocketstack-matt @Thels
 
+/shared/ @aidanm3341 @lbulanti-ms @willosborne @grahampacker-ms @jpgough-ms @rocketstack-matt @Thels
+
 /spectral/ @willosborne @lbulanti-ms @grahampacker-ms @jpgough-ms @rocketstack-matt
 
 /translator/ @Budlee @matthewgardner @jpgough-ms @rocketstack-matt
