+++ .gitignore
@@ -18,4 +18,7 @@ website/package-lock.json
 **/.vscode/
 
 node_modules/
+**/dist/
+**/.vscode-test/
+**/out/
 **/dependency-check-report/
