+++ cli/tsconfig.json
@@ -1,13 +1,15 @@
 {
   "compilerOptions": {
-    "module": "ES2020",
+    "outDir": "./dist",
+    "module": "CommonJS",
+    "target": "ES6",
     "esModuleInterop": true,
-    "target": "ES2021",
-    "moduleResolution": "node",
-    "outDir": "dist",
+    "forceConsistentCasingInFileNames": true,
+    "strict": true,
+    "skipLibCheck": true,
     "resolveJsonModule": true
   },
-  "lib": [
-    "es2021"
+  "include": [
+    "src/**/*"
   ]
 }
