+++ shared/README.md
@@ -1,0 +1,3 @@
+# Shared Logic Module
+
+This module provides shared logic such as validation and visualization utilities, intended for use across various plugins and tools in the codebase. It simplifies code reuse and promotes a unified logic layer, making it easier to maintain and extend.
