+++ shared/package.json
@@ -1,0 +1,76 @@
+{
+    "name": "@finos/calm-shared",
+    "version": "0.2.1",
+    "description": "A set of tools for interacting with the Common Architecture Language Model (CALM)",
+    "main": "dist/index.js",
+    "types": "dist/index.d.ts",
+    "files": [
+        "dist/"
+    ],
+    "private": true,
+    "scripts": {
+        "build": "tsc && npm run copy-calm-schema && npm run copy-spectral-rules",
+        "watch": "tsc -watch -p ./",
+        "clean": "rimraf node_modules dist",
+        "test": "jest --verbose",
+        "lint": "eslint src --config ./.eslintrc.json",
+        "lint-fix": "eslint src -- --fix",
+        "copy-calm-schema": "copyfiles ../calm/draft/2024-04/meta/* dist/calm/",
+        "copy-spectral-rules": "copyfiles ../spectral/**/* dist/spectral/",
+        "copy-calm-schema-old": "mkdir -p dist/calm && cp -r ../calm/draft/2024-04/meta dist/calm/",
+        "dependency-check": "dependency-check --project 'calm-shared' --scan . --out ./dependency-check-report --format ALL --suppression ../.github/node-cve-ignore-list.xml"
+    },
+    "eslintConfig": {
+        "extends": "../.eslintrc.json"
+    },
+    "bin": {
+        "calm": "dist/index.js"
+    },
+    "keywords": [
+        "cli",
+        "validation"
+    ],
+    "author": "",
+    "license": "ISC",
+    "module": "commonjs",
+    "type": "commonjs",
+    "dependencies": {
+        "@stoplight/spectral-cli": "^6.13.1",
+        "@stoplight/spectral-core": "^1.19.1",
+        "@typescript-eslint/parser": "^8.9.0",
+        "@viz-js/viz": "^3.9.0",
+        "ajv": "^8.12.0",
+        "copyfiles": "^2.4.1",
+        "graphviz-cli": "^2.0.0",
+        "js-yaml": "^4.1.0",
+        "json-pointer": "^0.6.2",
+        "jsonpath-plus": "^10.0.0",
+        "junit-report-builder": "^5.0.0",
+        "lodash": "^4.17.21",
+        "mkdirp": "^3.0.1",
+        "ts-graphviz": "^2.1.1",
+        "winston": "^3.14.1"
+    },
+    "devDependencies": {
+        "@jest/globals": "^29.7.0",
+        "@stoplight/types": "^14.1.1",
+        "@types/jest": "^29.5.13",
+        "@types/js-yaml": "^4.0.9",
+        "@types/json-pointer": "^1.0.34",
+        "@types/junit-report-builder": "^3.0.2",
+        "@types/lodash": "^4.17.0",
+        "@typescript-eslint/eslint-plugin": "^8.0.0",
+        "eslint": "^8.57.0",
+        "fetch-mock": "^9.11.0",
+        "jest": "^29.7.0",
+        "rimraf": "^6.0.1",
+        "ts-jest": "^29.2.5",
+        "ts-node": "^10.9.2",
+        "typescript": "^5.4.3"
+    },
+    "overrides": {
+        "path-to-regexp": "3.3.0",
+        "rollup": "4.24.0",
+        "jsonpath-plus": "^10.0.0"
+    }
+}
