+++ shared/src/commands/generate/components/metadata.spec.ts
