+++ shared/src/commands/generate/components/metadata.ts
