+++ shared/src/commands/generate/components/node.spec.ts
