+++ shared/src/commands/generate/components/node.ts
