+++ shared/src/commands/generate/components/property.spec.ts
