+++ shared/src/commands/generate/components/property.ts
