+++ shared/src/commands/generate/components/relationship.spec.ts
