+++ shared/src/commands/generate/components/relationship.ts
