+++ shared/src/commands/generate/generate.e2e.spec.ts
