+++ shared/src/commands/generate/generate.spec.ts
