+++ shared/src/commands/generate/generate.ts
