+++ shared/src/commands/generate/schema-directory.spec.ts
