+++ shared/src/commands/generate/schema-directory.ts
