+++ shared/src/commands/generate/util.ts
