+++ shared/src/commands/helper.ts
