+++ shared/src/commands/validate/spectral.result.ts
