+++ shared/src/consts.ts
@@ -1,0 +1,2 @@
+export const CALM_META_SCHEMA_DIRECTORY = __dirname + '/calm/meta';
+export const CALM_SPECTRAL_RULES_DIRECTORY = __dirname + '/spectral';
