+++ shared/test_fixtures/api-gateway-implementation-that-does-not-match-schema.json
