+++ shared/test_fixtures/api-gateway-implementation-that-does-not-pass-the-spectral-validation.json
