+++ shared/test_fixtures/api-gateway-implementation.json
