+++ shared/test_fixtures/api-gateway-self-reference.json
