+++ shared/test_fixtures/api-gateway-with-no-relationships.json
