+++ shared/test_fixtures/bad-schema/bad-json-schema.json
