+++ shared/test_fixtures/calm/core.json
