+++ shared/test_fixtures/core.json
