+++ shared/test_fixtures/recursive_refs/recursive.json
